Move editor utils functions out of api
