+++ spyder/utils/editor.py
