+++ spyder/utils/syntaxhighlighters.py
@@ -28,7 +28,7 @@ from spyder.config.base import _
 from spyder.config.main import CONF
 from spyder.py3compat import builtins, is_text_string, to_text_string
 from spyder.utils.sourcecode import CELL_LANGUAGES
-from spyder.api.utils import TextBlockHelper as tbh
+from spyder.utils.editor import TextBlockHelper as tbh
 
 PYGMENTS_REQVER = '>=2.0'
 dependencies.add("pygments", _("Syntax highlighting for Matlab, Julia and "
